Ensure that finish() means no more screenshots are taken
This fixes broken animated gifs I've seen where an additional image was
written on after the gif was 'finished'.

This was basically a race condition between the log-listener thread and
the main thread - the main thread will now block while waiting for
the last screenshot to be taken if it's already in progress, which works
much better.